fix(copy): reverted context code
